+++ code/src/index.js
@@ -1,10 +1,10 @@
-const cors = require("cors");
-const express = require("express");
-const routes = require("./routes/routes");
-const config = require("./settings/config");
+import cors from 'cors';
+import express from 'express';
+import routes from './routes/routes';
+import config from './settings/config';
+
 const port = config.NODE_PORT;
 const server = express();
-const pool = require("./database/connector");
 
 server.use(cors());
 server.use(express.json());
@@ -13,7 +13,7 @@ server.use(routes);
 server.listen(port, (error) => {
   if (error) {
     console.error(`Error starting Server ${error}`);
-  }else{
+  } else {
     console.log(`Server listening at port ${port}.`);
   }
 });